[API stabilization] Single entry point for serializers retrieval, Serializers.kt
    * Rename serializerByTypeToken to serializer
    * Internal bookkeeping, cleanup, leverage Kotlin 1.4 improvements
