+++ gradle.properties
@@ -36,3 +36,4 @@ kotlin.native.distribution.type=prebuilt
 # Workaround for Bintray treating .sha512 files as artifacts
 # https://github.com/gradle/gradle/issues/11412
 systemProp.org.gradle.internal.publish.checksums.insecure=true
+org.gradle.jvmargs="-XX:+HeapDumpOnOutOfMemoryError"
