--- runtime/jsMain/src/kotlinx/serialization/internal/Arrays.kt
@@ -1,15 +1,0 @@
-/*
- * Copyright 2017-2020 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
- */
-
-package kotlinx.serialization.internal
-
-internal actual fun <T> Array<T>.getChecked(index: Int): T {
-    if (index !in indices) throw IndexOutOfBoundsException("Index $index out of bounds $indices")
-    return get(index)
-}
-
-internal actual fun BooleanArray.getChecked(index: Int): Boolean {
-    if (index !in indices) throw IndexOutOfBoundsException("Index $index out of bounds $indices")
-    return get(index)
-}
