--- runtime/jvmMain/src/kotlinx/serialization/internal/Arrays.kt
@@ -1,15 +1,0 @@
-/*
- * Copyright 2017-2020 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
- */
-
-package kotlinx.serialization.internal
-
-@Suppress("NOTHING_TO_INLINE")
-internal actual inline fun <T> Array<T>.getChecked(index: Int): T {
-    return get(index)
-}
-
-@Suppress("NOTHING_TO_INLINE")
-internal actual inline fun BooleanArray.getChecked(index: Int): Boolean {
-    return get(index)
-}
